+++ client/src/App.js
@@ -13,8 +13,9 @@ import Auth from './utils/auth';
 /// IMPORT PAGES ///
 import Landing from './pages/Landing';
 import Signup from './pages/Signup';
-import Login from './pages/Login';
+import Login from './components/Login/index';
 import Dashboard from './pages/Dashboard';
+
 
 /// IMPORT COMPONENTS ///
 import ProtectRoute from './components/ProtectRoute';
